HADOOP_HOME handling and messaging
Update adds handling for hadoop.home.dir dflag as well as HADOOP_HOME.
This includes verification that one of the variables is set and that
the path exists, is a directory, and is an absolute path.  The reason
we verify these is that executables referenced in this path are
built-in to Hadoop and overriding them could defeat security measures.
To simplify unambiguous usage of these executables, I also added a
utility function -- getQualifiedBinPath that will return a fully
qualified path to a hadoop binary if it exists in $HADOOP_HOME\bin
or will throw.  In each case of its usage today, added clear error
propagation to indicate the underlying reason for the failure.