Patch for Classpath too long issue
Currently the classpath is set for Hadoop job tasks through �set CLASSPATH=..� in taskJvm.cmd. If the path gets too long we get an error that a command length exceeded the max limit of 8192 characters.

As we have integrated Oozie, and started running Oozie jobs, we hit errors in TT due to this issue. For example, as Oozie tries to start a hive job, it fails with the following error:
java.io.IOException: Command exceeds the OS command length limit: 8192, command: "set CLASSPATH="�.

The classpath includes Hadoop Core + Hive Jars which are in the distributed cache. This is causing the classpath to be too long giving the error above. A viable long term fix is to generate a temporary JAR file in the task cache directory which includes all elements in the long classpath, and then provide this as the �-classpath� argument for the JVM to be spawned. This is gated for Windows only.