CombineFileInputFormat only works with default file system
CombineFileInputFormat does not work with ASV. This appeared as an issue
in Sqoop which failed to export files in ASV. CombineFileInputFormat strips
out the scheme and authority components of the path, after which point the
path is assumed to be on the default file system.

Sqoop has it own copy of CombineFileInputFormat, but updated the ones in
core as well for consistency.

There are currently open Jiras for the same issue:
 https://issues.apache.org/jira/browse/MAPREDUCE-2704
 https://issues.apache.org/jira/browse/MAPREDUCE-1806

Attached a patch to use the full path, instead of only the path component
of the URI. Updated the tests to check for the expected perfix in the path
component of the URI rather than the full path. Verified core tests are
passing