HADOOP-90 Fix TestCommandLineJobSubmission
As we generate a Jar file to reference classpath entries in Windows, this test
started to fail since it expects to find testjob.jar and test.jar directly in
classpath property which contains only the referencing jar in case of Windows.

Changed the test to also look into the manifest of the referencing Jar if the
expected items were not found directly in classpath property.