PROD-82 - hadoop-env.cmd overwrites the value of *_OPTS set before install
The values of the following environment variables are overwritten in the hadoop-env.cmd file. This blocks us from using these variables for setting other properties. I am trying to set JMX properties for the hadoop processes but they don't take effect due to this.

This changeset appends the existing values if any.